Enhancements and additional functionality
Added a logging procedure to PacketHandler to wrap FLogger.
Added the total packet length to the packet handling functions
arguments.
Renamed CheckForError in TSnifferThread to CheckForPcapError.
Added handling support for IPv4, TCP and UDP.
General tidying up of code. Packet handling functions now get a pointer
pointing to their section of the packet (ie. IPv4 get a pointer that
already has the length of the Ethernet header added to it).
Added functions to handle byte ordering.
And correctly terminated FSnifferThread.